chore(cleanup)t: Refactor database utilities and reorganize feature modules
