+++ src/features/counterparty/counterparty_utils.py
