--- src/utils/data/database/__init__.py
@@ -1,1 +1,0 @@
-# This file makes the ai directory a Python package.
